pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darpaa-moe-intermediary-layer"
version = "0.1.0"
description = "DARPA-ready MoE intermediary layer"
authors = [{ name = "Autogenerated", email = "dev@example.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "redis>=5.0.4",
    "celery[redis]>=5.4.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "prometheus-client>=0.20.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "tenacity>=8.2.3",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.1",
    "orjson>=3.10.0",
    "numpy>=1.26.4",
    "scikit-learn>=1.4.2",
    "cryptography>=42.0.8",
    "structlog>=24.1.0",
    "itsdangerous>=2.2.0",
    "slowapi>=0.1.9",
    "sentence-transformers>=3.0.1",
    "whoosh>=2.7.4",
    "python-dateutil>=2.9.0.post0",
    "types-redis",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.1",
    "types-pyyaml",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
addopts = "-q --maxfail=1"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.cyclonedx]
output-format = "json"

